Revised code per comments
